refactor(winningNumber): 로또 번호 검증 비즈니스 로직 도메인 내부로 변경
Before:

scope: {
- winningNumberService: 로또 번호 검증 로직
- Lotto를 가지고 검증을 한다
}

After: {
- winningNumber 내부에서 로또 번호 검증 로직을 가진다
- numbers만 받아서 검증을 한다
}
