+++ src/test/java/lotto/winningNumber/service/WinningNumberServiceTest.java
@@ -32,50 +32,6 @@ class WinningNumberServiceTest {
                 .containsExactlyInAnyOrder(List.of(1, 2, 3, 4, 5, 6), 7);
     }
 
-    @ParameterizedTest
-    @DisplayName("로또와 당첨번호를 받아 로또의 당첨개수를 센다")
-    @MethodSource("providedMatchCount")
-    void calculateMatchCount(List<Integer> lottoNumbers, List<Integer> winningNumbers,
-                             int bonus, int count) throws Exception {
-        // given
-        Lotto lotto = new Lotto(lottoNumbers);
-        WinningNumber winningNumber = new WinningNumber(winningNumbers, bonus);
 
-        // when
-        int result = winningNumberService.calculateMatchCount(lotto, winningNumber);
-
-        // then
-        assertThat(result).isEqualTo(count);
-    }
-
-    private static Stream<Arguments> providedMatchCount() {
-        return Stream.of(
-                Arguments.arguments(List.of(1, 2, 3, 4, 5, 6), List.of(1, 2, 3, 4, 5, 6), 7, 6),
-                Arguments.arguments(List.of(42, 32, 12, 22, 45, 2), List.of(13, 24, 24, 22, 32, 1), 2, 2)
-        );
-    }
-
-    @ParameterizedTest
-    @DisplayName("로또와 당첨번호를 받아 보너스가 맞았는지 확인한다")
-    @MethodSource("providedMatchedBonus")
-    void isBonusMatched(List<Integer> lottoNumbers, List<Integer> winningNumbers,
-                             int bonus, boolean matchedBonus) throws Exception {
-        // given
-        Lotto lotto = new Lotto(lottoNumbers);
-        WinningNumber winningNumber = new WinningNumber(winningNumbers, bonus);
-
-        // when
-        boolean result = winningNumberService.isBonusMatched(lotto, winningNumber);
-
-        // then
-        assertThat(result).isEqualTo(matchedBonus);
-    }
-
-    private static Stream<Arguments> providedMatchedBonus() {
-        return Stream.of(
-                Arguments.arguments(List.of(1, 2, 3, 4, 5, 6), List.of(1, 2, 3, 4, 5, 6), 7, false),
-                Arguments.arguments(List.of(42, 32, 12, 22, 45, 2), List.of(13, 24, 24, 22, 32, 1), 2, true)
-        );
-    }
 
 }
